refactor, remove BS dependency
