+++ package.json
@@ -83,7 +83,6 @@
     "ramda": "^0.27.0",
     "react": "^17.0.0",
     "react-beautiful-dnd": "^13.0.0",
-    "react-bootstrap": "^0.32.0",
     "react-diff-viewer": "^3.0.0",
     "react-dnd": "^14.0.0",
     "react-dnd-html5-backend": "^14.0.0",
@@ -129,7 +128,6 @@
     "@types/ramda": "^0.27.0",
     "@types/react": "^17.0.0",
     "@types/react-beautiful-dnd": "^13.0.0",
-    "@types/react-bootstrap": "^0.32.0",
     "@types/react-dom": "^17.0.0",
     "@types/react-lazyload": "^3.0.0",
     "@types/react-modal": "^3.0.0",
