+++ src/components/edtr-io/plugins/types/course.tsx
@@ -12,7 +12,8 @@ import {
   OptionalChild,
   entityType,
 } from './common'
-import { RevisionHistory, Settings } from './helpers/settings'
+import { RevisionHistory } from './helpers/revision-history'
+import { Settings } from './helpers/settings'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
 
 export const courseTypeState = entityType(
