+++ src/components/edtr-io/plugins/types/helpers/settings.tsx
@@ -1,162 +1,10 @@
 /* eslint-disable import/no-internal-modules */
 
-import { PluginToolbarButton } from '@edtr-io/core'
 import { StateTypeReturnType, string } from '@edtr-io/plugin'
-import { Icon, faCheck, styled } from '@edtr-io/ui'
-import { faHistory } from '@fortawesome/free-solid-svg-icons/faHistory'
-import moment from 'moment'
-import * as React from 'react'
-import BSButton from 'react-bootstrap/lib/Button'
-import BSControlLabel from 'react-bootstrap/lib/ControlLabel'
-import BSFormControl from 'react-bootstrap/lib/FormControl'
-import BSFormGroup from 'react-bootstrap/lib/FormGroup'
-import BSModal from 'react-bootstrap/lib/Modal'
-import BSTable from 'react-bootstrap/lib/Table'
+import clsx from 'clsx'
+import { PropsWithChildren } from 'react'
 
-import { useLoggedInData } from '@/contexts/logged-in-data-context'
-
-const StyledTR = styled.tr<{ selected: boolean }>((props) => {
-  return props.selected
-    ? {
-        border: '3px solid rgb(0,100,0)',
-      }
-    : {
-        cursor: 'pointer',
-      }
-})
-
-interface RevisionData {
-  id: number
-  timestamp: string
-  author: string
-  changes: string
-  active: boolean
-}
-
-export function RevisionHistory<T>(
-  props: React.PropsWithChildren<{
-    id: number
-    currentRevision: number
-    onSwitchRevision: (data: T) => void
-  }>
-) {
-  const [availableRevisions] = React.useState<RevisionData[]>([])
-  const [showRevisions, setShowRevisions] = React.useState(false)
-  React.useEffect(() => {
-    if (props.id !== 0) {
-      /*fetch(`https://de.serlo.org/entity/repository/get-revisions/${props.id}`)
-        .then((response) => response.json())
-        .then((data: RevisionData[]) => {
-          setAvailableRevisions(data)
-        })*/
-    }
-  }, [props.id])
-
-  const loggedInData = useLoggedInData()
-  if (!loggedInData) return null
-  const editorStrings = loggedInData.strings.editor
-
-  return (
-    <div>
-      <span
-        onClick={() => {
-          if (availableRevisions.length) {
-            setShowRevisions(true)
-          }
-        }}
-      >
-        <PluginToolbarButton
-          icon={<Icon icon={faHistory} size="lg" />}
-          label={editorStrings.edtrIo.switchRevision}
-        />
-      </span>
-      <BSModal
-        show={showRevisions}
-        onHide={() => {
-          setShowRevisions(false)
-        }}
-        bsSize="lg"
-      >
-        <BSModal.Header closeButton>
-          <BSModal.Title>{editorStrings.edtrIo.switchRevision}</BSModal.Title>
-        </BSModal.Header>
-        <BSModal.Body>
-          <BSTable striped hover>
-            <thead>
-              <tr>
-                <th>#</th>
-                <th>{editorStrings.edtrIo.current}</th>
-                <th>{editorStrings.edtrIo.changes}</th>
-                <th>{editorStrings.edtrIo.author}</th>
-                <th>{editorStrings.edtrIo.createdAt}</th>
-              </tr>
-            </thead>
-            <tbody>
-              {availableRevisions.map((revisionData) => {
-                const selected = props.currentRevision
-                  ? props.currentRevision === revisionData.id
-                  : revisionData.active
-
-                const dateTime = moment.utc(revisionData.timestamp).local()
-                return (
-                  <StyledTR
-                    selected={selected}
-                    onClick={() => {
-                      // don't select the current selected
-                      if (selected) return
-                      // TODO: Build own fetch
-                      alert('not implemented yet')
-                      /*void fetch(
-                        `/entity/repository/get-revision-data/${props.id}/${revisionData.id}`
-                      )
-                        .then((response) => response.json())
-                        .then((data: { state: unknown; type: string }) => {
-                          const deserialized = deserialize({
-                            initialState: data.state,
-                            type: data.type,
-                          })
-                          if (isError(deserialized)) {
-                            alert(deserialized.error)
-                          } else {
-                            props.onSwitchRevision(
-                              deserialized.initialState.state as T
-                            )
-                            setShowRevisions(false)
-                          }
-                        })*/
-                    }}
-                    key={revisionData.id}
-                  >
-                    <td>{revisionData.id}</td>
-                    <td>
-                      {revisionData.active ? <Icon icon={faCheck} /> : null}
-                    </td>
-                    <th>{revisionData.changes}</th>
-                    <td>{revisionData.author}</td>
-                    <td title={dateTime.format('LL, LTS')}>
-                      {dateTime.fromNow()}
-                    </td>
-                  </StyledTR>
-                )
-              })}
-            </tbody>
-          </BSTable>
-        </BSModal.Body>
-        <BSModal.Footer>
-          <BSButton
-            onClick={() => {
-              setShowRevisions(false)
-            }}
-          >
-            Schließen
-          </BSButton>
-        </BSModal.Footer>
-      </BSModal>
-    </div>
-  )
-}
-
-export function Settings(props: React.PropsWithChildren<{}>) {
+export function Settings(props: PropsWithChildren<{}>) {
   return <>{props.children}</>
 }
 
@@ -168,17 +16,20 @@ Settings.Textarea = function SettingsTextarea({
   state: StateTypeReturnType<ReturnType<typeof string>>
 }) {
   return (
-    <BSFormGroup>
-      <BSControlLabel>{label}</BSControlLabel>
-      <BSFormControl
-        componentClass="textarea"
+    <label className="font-bold">
+      {label}
+      <textarea
         value={state.value}
         onChange={(e) => {
           const { value } = e.target as HTMLTextAreaElement
           state.set(value)
         }}
+        className={clsx(
+          'mt-1 mb-7 flex items-center rounded-2xl w-full p-2',
+          'bg-brand-150 border-2 border-brand-150 focus-within:outline-none focus-within:border-brand-light'
+        )}
       />
-    </BSFormGroup>
+    </label>
   )
 }
 
@@ -192,25 +43,27 @@ Settings.Select = function SettingsSelect({
   options: { label: string; value: string }[]
 }) {
   return (
-    <BSFormGroup controlId="formControlsSelect">
-      <BSControlLabel>{label}</BSControlLabel>
-      <BSFormControl
-        componentClass="select"
-        placeholder="select"
-        value={state.value}
-        onChange={(e) => {
-          const { value } = e.target as HTMLSelectElement
-          state.set(value)
-        }}
-      >
-        {options.map((option) => {
-          return (
-            <option key={option.value} value={option.value}>
-              {option.label}
-            </option>
-          )
-        })}
-      </BSFormControl>
-    </BSFormGroup>
+    <div>
+      <label className="font-bold">
+        {label}
+        {/* TODO: find example, add styling */}
+        <select
+          placeholder="select"
+          value={state.value}
+          onChange={(e) => {
+            const { value } = e.target as HTMLSelectElement
+            state.set(value)
+          }}
+        >
+          {options.map((option) => {
+            return (
+              <option key={option.value} value={option.value}>
+                {option.label}
+              </option>
+            )
+          })}
+        </select>
+      </label>
+    </div>
   )
 }
