+++ src/components/edtr-io/plugins/types/video.tsx
@@ -9,7 +9,7 @@ import {
   HeaderInput,
   entityType,
 } from './common'
-import { RevisionHistory } from './helpers/settings'
+import { RevisionHistory } from './helpers/revision-history'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
 
 export const videoTypeState = entityType(
